Windows Python 3.7 3.8 and Boost Compatibility Err
Because of changes to standard library interface on windows, Python 3.7 and 3.8 cause error with Boost headers.

Python 3.7 and 3.8 headers redefine sprintf as _sprintf. When Boost attempts to use sprintf, it is redirected to _sprintf, which was reimplemented as sprintf and does not exist. See https://github.com/boostorg/system/issues/32.

Resolution: we are deprecating the support for Python 3.7 and 3.8 on Windows as the error cannot be resolved without making changes to the Python headers.